requirements.txt
# Core
confluent-kafka
pandas
pyarrow
psycopg2-binary

# Spark
pyspark

# ML
scikit-learn
joblib

# Visualization
streamlit
plotly

# Utils
python-dotenv
loguru

# Notebooks
jupyter
ipykernel
